Refactor index.ts
